+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.34.1] - 2026-04-07
+### Changed
+- Renamed `component_url` field to `url` in `ComponentLicenseInfo` message for consistency with other URL fields in the API
 
 ## [0.34.0] - 2026-04-01
 ### Added
@@ -262,6 +265,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Vulnerabilities
 - Added REST endpoint support for each service also
 
+[0.34.1]: https://github.com/scanoss/papi/compare/v0.34.0...v0.34.1
 [0.34.0]: https://github.com/scanoss/papi/compare/v0.33.0...v0.34.0
 [0.33.0]: https://github.com/scanoss/papi/compare/v0.32.1...v0.33.0
 [0.32.1]: https://github.com/scanoss/papi/compare/v0.32.0...v0.32.1